+++ tensorcircuit/backends/numpy_backend.py
@@ -91,6 +91,40 @@ class NumpyBackend(numpy_backend.NumPyBackend):  # type: ignore
     def cos(self, a: Tensor) -> Tensor:
         return np.cos(a)
 
+    # acos acosh asin asinh atan atan2 atanh cosh (cos) tan tanh sinh (sin)
+    def acos(self, a: Tensor) -> Tensor:
+        return np.arccos(a)
+
+    def acosh(self, a: Tensor) -> Tensor:
+        return np.arccosh(a)
+
+    def asin(self, a: Tensor) -> Tensor:
+        return np.arcsin(a)
+
+    def asinh(self, a: Tensor) -> Tensor:
+        return np.arcsinh(a)
+
+    def atan(self, a: Tensor) -> Tensor:
+        return np.arctan(a)
+
+    def atan2(self, y: Tensor, x: Tensor) -> Tensor:
+        return np.arctan2(y, x)
+
+    def atanh(self, a: Tensor) -> Tensor:
+        return np.arctanh(a)
+
+    def cosh(self, a: Tensor) -> Tensor:
+        return np.cosh(a)
+
+    def tan(self, a: Tensor) -> Tensor:
+        return np.tan(a)
+
+    def tanh(self, a: Tensor) -> Tensor:
+        return np.tanh(a)
+
+    def sinh(self, a: Tensor) -> Tensor:
+        return np.sinh(a)
+
     def size(self, a: Tensor) -> Tensor:
         return a.size
 
@@ -115,6 +149,14 @@ class NumpyBackend(numpy_backend.NumPyBackend):  # type: ignore
 
     def tile(self, a: Tensor, rep: Tensor) -> Tensor:
         return np.tile(a, rep)
+
+    def mean(
+        self,
+        a: Tensor,
+        axis: Optional[Sequence[int]] = None,
+        keepdims: bool = False,
+    ) -> Tensor:
+        return np.mean(a, axis=axis, keepdims=keepdims)
 
     def unique_with_counts(self, a: Tensor) -> Tuple[Tensor, Tensor]:
         return np.unique(a, return_counts=True)  # type: ignore